Run Engine tests in dev Engine (on top of stable Engine)
So that we can find potential failures earlier, as we've learned from
v0.8.0 & v0.8.3 releases.

Allow PRs to use both Dagger runners & paid GitHub runners (same size).
It's OK if these fail for now - `continue-on-error` - since we seem to
have a configuration issue on the self-hosted runners themselves (they
are attempting to use the same Engine). We can follow-up with a fix.

By the way, if we like this change, we can roll it out to other jobs
- supersedes https://github.com/dagger/dagger/pull/5359

Stop uploading journal.log. Since we are reusing workflows, we cannot
easily determine the journal log file name (the same one gets
overwritten by different workflows). If we use inputs.mage-targets to
distinguish, this fails since it contains ':', an unsupported character
by the upload-artifact action. We could do this better (e.g. by
requiring another input, by handling this in code, etc.), but I am
unsure what is the current value of journal.log when we use Dagger
Cloud. Removing it for now, let's see if others miss it.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>