fix: calling id for any optional arguments that is ID type
Some functions, it has optional argument that is ID type. Before this
change, if we calling that function, it will crash because that function
didn't calling `id/1` to convert into ID string. This change fix that
issue by calling `id/1` or iterate calling `id/1` for any function that
accept list of ID type, such as, `Dagger.Directory.docker_build/2`.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>