Add stop in Python
The Python SDK has multiple types of connections which are decoupled from the client instance. It would be more natural to call `client.stop()` when automatically closing that connection but that's in a lower level.

Both `dagger.Connection` (legacy) and `dagger.connection` (experimental) support automatic provisioning, so pushed the common client setup to the engine connection manager.

The connect timeout was documented as 10.0 but set as 30.0 by mistake during development of the default client. Using the same value for stop to avoid adding more settings. It could be better documented but the timeout setting is in the middle of a change in format (see Config.__post_init__). This can all be revisited later.

To simplify the use case of codegen starting out with a blank `gen.py` file, it's now required to run under an existing session like `dagger run`. Note that this type of connection doesn't setup a client and so it doesn't add the `client.stop` as well.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>