Backport: services via gateway, minus API changes (#5764)
* TestPlatformCrossCompile: avoid panic in goroutine

Signed-off-by: Alex Suraci <alex@dagger.io>

* buildkit.Client: tweaks for more resiliency

Ran into a few panics/cryptic errors here while switching services over
to the gateway. I don't remember exactly if these were fixed (a lot
happened between now and then) but it seems worth committing anyway.

Also, change ResolveImageConfig to be a passthru along the way, for no
real reason.

Signed-off-by: Alex Suraci <alex@dagger.io>

* run services via gateway

* Service is now a distinct type, but only used internally
* Services now run via the gateway interface instead
* (core.Container).Service returns a Service that runs the container
* Services can be explicitly started and stopped (not exposed yet)
* refactor ServiceBindings, merge with HostAliases
  * embed a Service instead of a ServiceID

Signed-off-by: Alex Suraci <alex@dagger.io>

* de-integrate Services from API, remove c2h/h2c

making this a separate commit just in case it makes it easier to bring
these changes back in the future.

part of this is also just adding files that I accidentally removed (via
git clean) while making the previous commit.

Signed-off-by: Alex Suraci <alex@dagger.io>

* add Query { stop }

Signed-off-by: Alex Suraci <alex@dagger.io>

* regen SDKs

Signed-off-by: Alex Suraci <alex@dagger.io>

* sdk/go: stop on client close

Signed-off-by: Alex Suraci <alex@dagger.io>

* sdk/nodejs: fix some editor lints

Signed-off-by: Alex Suraci <alex@dagger.io>

* Query.stop: support timeout

gives meaning for the boolean return value, makes it easier for clients
to support timeouts

Signed-off-by: Alex Suraci <alex@dagger.io>

* attempt updating elixir SDK

Signed-off-by: Alex Suraci <alex@dagger.io>

* sdk/go, sdk/nodejs: 10 second stop timeout

Signed-off-by: Alex Suraci <alex@dagger.io>

* sdk/nodejs: fix stopping (use finally)

Signed-off-by: Alex Suraci <alex@dagger.io>

* regen python

Signed-off-by: Alex Suraci <alex@dagger.io>

* sdk/python: stop on client close

Signed-off-by: Alex Suraci <alex@dagger.io>

* sdk/python: test config, stop with global client

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert changes to connection in Python

This reverts commits c142c49 and 51429e5.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Add stop in Python

The Python SDK has multiple types of connections which are decoupled from the client instance. It would be more natural to call `client.stop()` when automatically closing that connection but that's in a lower level.

Both `dagger.Connection` (legacy) and `dagger.connection` (experimental) support automatic provisioning, so pushed the common client setup to the engine connection manager.

The connect timeout was documented as 10.0 but set as 30.0 by mistake during development of the default client. Using the same value for stop to avoid adding more settings. It could be better documented but the timeout setting is in the middle of a change in format (see Config.__post_init__). This can all be revisited later.

To simplify the use case of codegen starting out with a blank `gen.py` file, it's now required to run under an existing session like `dagger run`. Note that this type of connection doesn't setup a client and so it doesn't add the `client.stop` as well.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* switch from GraphQL to internal API for shutdown

Signed-off-by: Alex Suraci <alex@dagger.io>

* undo SDK changes; no longer needed

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix shutdown deadlock, tidy up logs

Signed-off-by: Alex Suraci <alex@dagger.io>

* refactor tricky service synchronization

* Get: fix missing Unlock
* Start: use a good old loop + label to fix case where multiple Starts
  are waiting, and the first Start fails.

Signed-off-by: Alex Suraci <alex@dagger.io>

* refactor Services.Get

use a loop; if multiple start attempts overlap we can just keep waiting

Signed-off-by: Alex Suraci <alex@dagger.io>

* add unit tests for Services.Start

also remove Service from RunningService; not needed

Signed-off-by: Alex Suraci <alex@dagger.io>

* test starting same service in different clients

refactor stub service along the way

Signed-off-by: Alex Suraci <alex@dagger.io>

* minor cleanups

Signed-off-by: Alex Suraci <alex@dagger.io>

* add changelog

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>
Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>
Co-authored-by: Helder Correia <174525+helderco@users.noreply.github.com>