fix: attempt to start engine container before return
Occasionally, the engine container can end up in an inconsistent state,
and exist, but not be running. Previously, in this scenario the dagger
client would fail to create this container (since it already exists),
but it would not be started - this results in the client hanging
indefinitely.

To resolve this, we add new logic that detects when the engine container
with the target name already exists, and just attempt to restart it
instead of creating a new container.

Signed-off-by: Justin Chadwell <me@jedevc.com>