engine: support filesync w/ engine container; container asTarball API
This change enables us to interact with engine container's filesystem
the same way we do client filesystems in terms of file import+export.

The motivation was primarily to support some upcoming requirements
around SDK bootstrapping where builtin SDKs will be pre-loaded into the
engine container, which allows module usage w/out a hard dependency on
the internet (at least when using builtin module SDKs).

This also includes a new user-facing API that enables converting a
Container object to a File representation of it as an exported tarball.
This API will be very convenient to have for the upcoming work and be
useful for users in general. It saves you from having to export a
tarball to your local filesystem and then load it back into the engine,
instead doing that whole process internally and more efficiently by not
having to cross the wire all the way to the client and back.

That API is also helpful as some initial validation that the engine
container filesync is working as expected.

Signed-off-by: Erik Sipsma <erik@dagger.io>