assert on more specific message in internal pipeline test
This is a bit silly. The previous update to log otel failures caused the
word "merge" to appear in the output (it's an error about merging otel
schemas), which in turn caused our test assertion to fail because it's
checking that no "merge" (-op) vertexes show up in the output by
default.

There's not really any great way to make the test more robust than I can
see other than just asserting on a more specific string, "merge (". It's
important to have this test coverage though, so this seems like the best
solution.

Signed-off-by: Erik Sipsma <erik@dagger.io>