feat: allow querying gitref resolved commit
This adds a new commit resolver to a GitRef to get the resolved commit
id. This can be obtained from the provenance of the result after
evaluating it. There should no additional cost to this, since as long as
the parameters are the same, the cache should kick in, and we don't end
up evaluating a single git repository multiple times.

Signed-off-by: Justin Chadwell <me@jedevc.com>