engine: reduce connection retry noise
The engine client initializes its session w/ the server in a separate
goroutine (due to the fact that it blocks for the duration of the
session). This causes initial connection attempts to usually fail while
the session is still be setup.

This creates a lot of scary-looking but harmless logs in the common
case.

This changes the logic to only start printing failed connection attempts
once the retry logic is going to wait over a second before the next
attempt. The idea is to balance not showing useless errors but also not
leave the user hanging for too long with no output and not hide errors
that may useful for debugging.

Would love to have a test for this, but very hard to test timing things
w/out flakes, so just manually tested so far.

One potentially better solution would be to have a debug mode and always
hide errors unless that's enabled, but that would likely impact every
SDK because they all use the CLI via the `session` subcommand, so
scoping that out for now to get this fix in quick.

Signed-off-by: Erik Sipsma <erik@dagger.io>