core: namespace module objects and function extensions
This change results in names being prefixed by module name for:
1. Any objects defined by the module
2. Any functions defined by the module that are extending objects from
   other pre-existing modules (or core)

This reduces the chance of conflicts when depending on 2+ modules that
happen to use overlapping names.

It's implemented so that SDKs don't need any awareness of it; when a
runtime is called names are "de-namespaced" as needed so the runtime
receives names the same as what it originally defined.

Signed-off-by: Erik Sipsma <erik@dagger.io>