ci: remove conflicting cache mounts from rust test
We run CI on dagger runners now, which allows for much greater sharing
of cache mounts between different runs in parallel.

I saw some fairly consistent failures in the Rust SDK tests which seemed
to be related to missing files from cache mounts.

The cache mounts in question did not seem safe to be shared across
pipelines since they relied on the existence of a single file (as
opposed to something like cargo cache which I understand to be more like
a content addressed store or sorts).

This probably wasn't safe before either since cache mounts can be pruned
at any moment, so it's not safe to assume their contents will be present
in later execs.

Fixed this by removing some problematic cache mounts that couldn't be
namespaced and by namespacing the cargo cache mount with rust version
(in order to prevent conflicts between different versions of rust
running in parallel on the same dagger runner).

Signed-off-by: Erik Sipsma <erik@dagger.io>