fix: return upstream ErrNotFound from secret store
Having duplicate error types meant that the git llb source attempting to
look for the git secrets would not get an ErrNotFound, and would instead
bail after looking for the first secret.

This update fixes this by ensuring that we use the same error type, so
we properly look for all specified secrets.

Also adds some tests to verify the behaviour.

Signed-off-by: Justin Chadwell <me@jedevc.com>