core: disable support for modules extending core types
We've decided to disable this ability for now since it creates hard to
solve problems w/ namespacing (simply prefixing extension functions with
Module name often creates extremely awkward names) and is unclear in
terms of benefits relative to confusion in terms of API design.

It will be possible to re-add support for this in the future if desired
and in the meantime it would still be possible for a custom SDK to
implement support for this by identifying fields in the graphql schema
that could be implemented as methods directly attached to core types.

Signed-off-by: Erik Sipsma <erik@dagger.io>