chore: build buildctl from upstream
Previously, we were just grabbing it from an image, but now there's a
known vulnerability in grpc (which only affects the server, but oh well,
scanners are angry). To resolve this, we can build buildctl ourselves
from upstream, but bump grpc to v1.56.3 (since it's closest to the
v1.53.0 that buildkit is currently using, and there's an upstream PR to
handle this).

Signed-off-by: Justin Chadwell <me@jedevc.com>