fix: apply recursion only on added types in go SDK
The previous fix to apply recursion to detect nested types (in struct
fields, return parameters, etc) was too extensive.

Specifically: if a struct type X was created that contains struct type
Y, code for type Y would always be generated, even if type X would not
be generated (which could happen it was at the top-level and had no
functions declared on it).

To avoid this, we need to *only* consider any subtypes for code
generation iff the main type would be considered for code generation.
Doing this is a bit tricky though, we need to remove the global
extraTypes state, and instead return results back up through the
function call chain - this gives us a number of subtypes for each
struct, which we can conditionally add only if we've considered the
struct for codegen.

Signed-off-by: Justin Chadwell <me@jedevc.com>