zenith: always run git exec ops when resolving refs
Otherwise we can get stale data.

Thankfully the new `commit` API helps us with half of this.

Continuing to use an exec-op to get the default branch is far from
ideal though. The much better fix would be to move this all fully
into our core api so we don't need any exec ops. But this will
alleviate short term pain in the meantime.

Signed-off-by: Erik Sipsma <erik@dagger.io>