Move python module codegen introspection to cli
With the introspection we avoid having to make a connection to the API at all, which also avoids any issue importing an incomplete or non-importable `dagger` package (due to errors).

Cleaned up the python cli in preparation for moving the `mod init` templates there. Typer would have been a better choice for sys.argv parsing but I still wanted to avoid adding a dependency just for this.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>