ci: Add DAGGER_CLOUD_TOKEN client-side
Without this, even though the Engine already has this env var, runs will
**NOT** appear in https://dagger.cloud

To test the above, I have not yet configured this secret in the
repository. I will do a follow-up commit after the secret is set. If the
above is correct, the run for the first commit will not show in
https://dagger.cloud

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>