ci: Remove _EXPERIMENTAL_DAGGER_CLOUD_TOKEN from free GitHub runners
This is a legacy token that was replaced with DAGGER_CLOUD_TOKEN which
should be kept secret, meaning that it will not work in PRs or forks, so
there is no point in setting it.

The run for this commit will show in https://dagger.cloud since the
secret has been set in the repository config.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>