ci: add some primitive tests for daggerverse modules
Long-term goal is to be able to know if we are breaking existing
daggerverse modules as we make changes in PRs. Short-term goal is to
know if we are breaking existing daggerverse modules being used in
KubeCon.

This is incredibly barebones at the moment; just a hardcoded list of
modules w/out any pinning and a test case that calls `dagger functions`
on all of them in parallel in subtests. The hope is it's still useful
and potentially the start of something more fleshed out over the
long-term.

Signed-off-by: Erik Sipsma <erik@dagger.io>