fix: split out docker pull step in progress
This patch splits out the pull step to a separate place in the progress
output, which is only printed if the image is actually pulled from the
registry. This is added so we can have better insight into *what* part
of connecting to the dagger engine is actually taking time (currently
this is very opaque).

The first simple step to this is to split out the pull step, since this
is often the very slow part. Additionally, we make "starting engine"
include the "docker run" and the waiting for the client to be ready.

To actually do this, we need to do a fairly reasonable amount of
refactoring so that we don't end up adding too much new clutter to the
progress output. We only want to add the pulling output when necessary,
but it shouldn't be the child of "starting engine", that step should
come after.

Signed-off-by: Justin Chadwell <me@jedevc.com>