fix: allow nested structs from module functions
Previously this would just fail (but would allow nesting lists of
objects!). This was because the condition we were checking didn't take
into account non-idable object types (which is what's returned from
modules).

Also, this prevents the previously weird case of allowing a module to
declare it returns a directory and then return a *different* core type,
like a container, which would actually not error immediately.

Signed-off-by: Justin Chadwell <me@jedevc.com>