feat: make module types id-able
This feature allows all module declared objects to be converted to a
respective ID-type using the "id" field, and converted from an ID-type
using the "loadFromID" helper. This effectively makes these types
fully idable.

IDs are essentially the contents of the module type's data,
JSON-encoded, and then base64-encoded, similar (but not identical) to
standard IDs. At the moment, they don't share exactly the same codepaths
to normal modules, but this could be adapted to be so in the future. We
also include a note of *what type* generated this ID - this is so we can
prevent type confusion.

Essentially, this enables a number of previously impossible use cases:

1. Returning a custom object from another module from a function

   - This case is the simplest - we just extend our conversion that
     unpacks all returned IDs from functions into their underlying types
     (so while a module returns an XID, it can be queried as if it was X).

2. Accepting a custom object as a parameter to a function

   - Because graphql doesn't support creating inline objects, we get
     around this in the core api by instead passing IDs everywhere. All
     the code-gen is setup to expect this, but module types weren't
     IDable, so we failed during schema-checking.

     Additionally, we need to transparently convert the ID from the
     graphql request into the underlying storage data for the module.

3. Accepting a custom object from another module as a parameter to a
   function

   - This is the same case as above, but also involves ensuring that we
     stitch in the type declarations from the module into our own schema
     so that we have the right type.

     However, here, we **don't** decode the ID into its underlying
     data, it's a foreign type, so the module should only see the ID.

Signed-off-by: Justin Chadwell <me@jedevc.com>