engine: use our own dial-stdio binary rather than buildctl
The buildctl binary is currently required in the engine image since it
facilitates connhelpers connecting to the engine inside (by proxying the
stdio of an exec'd buildctl process to the unix socket). This is
inherited from buildkit and how Dagger has worked for a while.

The problem is we keep running into vulnerabilities in this binary that
take a while for upstream to fix (i.e. related to otel/grpc), which
leaves us to resort to various hacks to get a binary w/out
vulnerabilities in the engine image.

This change moves us off of upstream's buildctl binary and instead
replaces it with the absolute minimal features we need in
`./cmd/dialstdio`. There's none of the buildctl subcommands except
dial-stdio. This allows us to easily build it from our own go.mod (which
makes dependency upgrades easier) and also results in most problematic
sources of dependencies no longer being present.

The source code for the dial-stdio functionality is mostly a fork of
upstream, except all the other adjacent code we don't need has been
stripped out. I also switched to use cobra to be consistent w/ our other
binaries and avoid a separate dependency on urfave/cli.

Attribution on the original source code is given w/ a NOTICE and
BUILDKIT_AUTHORS file, same as we did w/ `./cmd/engine`.

Signed-off-by: Erik Sipsma <erik@dagger.io>