ci: invalidate apk install of engine image once per day.
We currently are in a situation where the engine image is being built
and installing apk packages that are vulnerable. Alpine released newer
versions of the packages a few days ago, but we are still hitting cache
for the "apk add" execs from before then and thus aren't picking them
up.

The real fix to this is to switch CI to use Zenith and use one of the
various apk/apko/wolfi modules that have been written which avoid this
problem, e.g.
https://daggerverse.dev/mod/github.com/sipsma/daggerverse/apk@d709dd631e176025ea042dc03fdc90197f4b12ba

However, in the meantime this change just wraps the apk add w/ an env
var that changes once per day (current unix time / 1 day in seconds),
which will hold us over until then.

Signed-off-by: Erik Sipsma <erik@dagger.io>