disallow use of external module types in function+field signatures
While we now give all module objects ID-ability and enable a module to
accept its own objects as function inputs, we are not yet going to allow
modules to use objects from other (non-core) modules as function args,
function returns or object fields.

This is because allowing that results in a caller who depends on a given
module A to now not only have access to module A's types but also the
types of transitive dependencies. This in turn opens the doors to
dependency hell; if two versions of a given module show up multiple
times in the full dependency DAG we need some fairly complex dependency
resolution logic around which of those versions is exposed to the
caller.

That problem is famously difficult and while we may need to deal with it
someday, it's preferable to avoid it for as long as possible.

The idea is to instead add support for interfaces in follow ups, which
should unblock many of the use cases that call for interoperability
between different module types.

The main changes in this commit are to detect when a user tries to do
this and give them a descriptive error message in that case. Other
related changes are:
* an additional integ test case for the situation where there's
  conflicting versions of a give module in the transitive dependency
  DAG, which is expected to continue working.
* tracking of which module a given schema type was loaded from, which
  allows us to distinguish between use of a type from core (which is
  still allowed) and a type from a dependency module.
* removal of "Dependencies" from the ExecutableSchema interface we use
  internally, which overlapped with some of the other changes here and
  was completely unused (leftover from ancient days during the Cloak
  project).

Signed-off-by: Erik Sipsma <erik@dagger.io>