core: move docker media type container import test off apko
apko no longer supports creating image tarballs with docker media types,
so we can't use it in our tests anymore.

It's nice to have coverage of importing a container tarball created by
something other than dagger, so I left apko in for the OCI part of the
container import test, but now we just use dagger to generate the docker
media type tarball.

To make up for some of the self-referentiality here I expanded the
coverage of TestContainerMediaTypes to verify that AsTarball also
respects the docker media types setting correctly.

Signed-off-by: Erik Sipsma <erik@dagger.io>