Update all website deps at once
Meant to address a bunch of CVEs.

Removing the yarn.lock and running `yarn install` again was the most
efficient way of getting this done. Everything seems to be working fine
locally, let's push this and see what happens on the preview
environment.

As a follow-up, we should upgrade to Docusaurus v3 so that we solve a
few more of the remaining CVEs:
https://docusaurus.io/docs/migration/v3

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>