feat: allow go modules to accept/return foreign types
This allows go modules to accept arguments and return values that are of
types from other packages that are not the current one.

For example, this allows module authors to structure their modules as
they like with sub-packages while not exposing those details to the
outside world.

This is a pretty straightforard change - to do this we need to:
- Disable the explicit check that disabled this behaviour
- Enabled packages.NeedDeps | packages.NeedImports to parse dependencies
  syntax so that we can analyze the struct and methods in those
  dependent packages.
- Update the logic when we reference type names to not always assume
  local, and to prefix the name of the package when we use it.
- Default to the main package when searching for syntax information, but
  fallback to the package that the type was declared in when we need to.

Signed-off-by: Justin Chadwell <me@jedevc.com>