fix: allow fast-cache hits on custom BlobSource
By prefixing the returned CacheKey with "session:" we activated the code
path in buildkit that randomizes the digest to ensure that we don't get
fast-cast hits (since session ids can be reused).

However, unlike LocalSources, BlobSources are not tied to a specific
session (after initializing them), so we actually want to avoid this
behavior. With this change, digests aren't randomized, so we can get the
fast-cache hits.

However, this isn't fully correct - see the comments for details.

Signed-off-by: Justin Chadwell <me@jedevc.com>