fix: only apply fancy pointer navigation for inline opts structs
This means that we will always pass in an inline opts struct without the
pointers, but we preserve the default value in the other cases.

So for example, given a `*string` parameter, if was optional and not
set, we should default to `nil`, not a pointer to an empty string (which
was the previous behaviour).

Signed-off-by: Justin Chadwell <me@jedevc.com>